Profile includes bugs

+++ src/settings.h
@@ -2,4 +2,4 @@
 #include <stdio.h>
 #include <stdlib.h> 
 //typedef double real_t;
-typedef size_t int_t;
+typedef unsigned int int_t;
